restructur renamed web to template
