+++ shared/db/authdb.go
