+++ shared/db/coursedb.go
