+++ shared/db/logdb.go
